Add mostly working happo integration
These things are now working:
- Grabbing config from .happo.js
- Create async report based on snap-requests created

I added a second spec file to make sure the full flow was working. I was
surprised to see that the before and after hooks became global for all
specs. They're attached to one spec in the logs but execute before and
after all tests have run in the entire test suite.