Fix makeAbsolute for relative urls
We weren't handling urls that are relative to the current document
location. e.g.

  node_modules/foo/bar.png

This example is from the cypress-example-todomvc app.

While we can't technically make these relative to the baseUrl, this is
at least better than failing. And I bet 95% of cases where people use
paths without a leading slash they actually meant to relate them to the
root of the application.