Add happo-cypress wrapper command
After a lot of debugging and exploration, I've landed in a fix for the
problem where a Cypress run creates multiple happo reports, one per test
file.

This was first reported last week and to begin with I couldn't repro it.
In short, this was the bug:

When you have multiple test files, Happo creates one report per file,
instead of one combined report.

When running with `cypress open`, the `after` hook was only called once
per full run. But if you ran `cypress run`, the behavior changed to
calling `after` once per test file. There's a bug for this in the
Cypress issue tracker: https://github.com/cypress-io/cypress/issues/5160

To mitigate this, I tried hard to find a place/hook or something where
Cypress would tell me one of these things:
The full suite is done, or
These are the specs in the full run

Either of those would be sufficient enough for our plugin to
conditionally create the Happo report once the `after` hook is called
for the last time. But I couldn't find any of this information available
from Cypress. So I did what Percy.io is doing -- added a wrapper command
that you need to use instead of `cypress run`:

happo-cypress -- cypress run

This call will set up a server locally, invoke the wrapped command,
accept `requestIds` to be POSTed to the server and then finally, once
the wrapped command is done, create the full Happo report.

It's a little unfortunate that we have to go down this route as there's
a lot of complexity involved:
- Users will have to make sure to wrap the Cypress command
- We have to be careful about choosing a good port (for now this is
hard-coded).