Resolve assets by document.location instead of `baseUrl`
When resolving assets (css files, images, etc), we were assuming that
they would come from `baseUrl` (as defined in cypress.json) when they
started with a `/`. That isn't true in case you're visiting pages that
aren't using the same baseUrl.

To fix this, I'm making all asset urls relative to the document they
were in when they were fetched. This removes the dependency between
happo-cypress and cypress.json as we can always grab the baseUrl from
the document.location.

Fixes #5