Add support for `happo-cypress finalize`
This command will be used in cases where you're running cypress tests in
parallel and you need to "finish up" a Happo report at the very end,
after all the test runs have finished.

An important part of this puzzle is the HAPPO_NONCE environment
variable which is a common identifier amongst all parallel runs. It
triggers slightly different behavior for the test run:
- individual `happo-cypress -- cypress run`s won't make comparisons
- each individual run will "patch" the async report with snap-request
ids