Inline canvases as PNG images
This commit will make sure that happo-cypress users can screenshoot
canvases. These are dynamically painted to, and there's no information
in the DOM about what they contain. So we take the canvas and convert it
to a PNG file.

For now, no styles from the canvas are transferred to the image. This
will mean that things like borders, margins, etc will go missing. I'm
going to address this in a follow-up. I might also address the fact that
we leave the dom "tainted". It might be best to clean up after ourselves
(putting back what we remove).