Clean up DOM after taking snapshot
Now that we replace canvases with images as part of the snapshot
process, it feels wrong to leave the replaced image in the DOM. While
this might not be an issue in many scenarios, it could pose problems to
things like React, where we'll confuse React's internal engine by
tampering with the DOM and not cleaning up.

We leave the cleanup to the inlineCanvases function, which can hold on
to references to elements that were replaced and make the opposite
replacement at the very end.