Make asset package smaller
Two fixes in this commit. The first one is that the same asset isn't
added multiple times to the zip file. This will make the payload
smaller.

The second fix is that the PNGs of inlined canvases are moved to the
asset package. This will ensure that they are deduped (via a hashing
mechanism to set the file name/url) which in turn makes the asset
package smaller.