Add `transformDOM` option
This feature can be used to replace certain DOM elements before they are
sent to happo for screenshooting. It is similar to Percy's
`domTransformation` but it has a slightly different signature. Since
we're transforming the actual DOM under test (similar to what we do for
canvases), we need a way to clean up after ourselves. For that reason,
the `selector` is part of the options passed to `transformDOM`.