Validate payload sent to happo-cypress server
We're seeing Percy post invalid (to us) payloads when Happo and Percy is
used alongside each other. To fix that, we can validate the payloads
before we act on them. We post lists of requestIds as numbers, separated
with newlines. So we can check to make sure we're only getting (valid)
numbers.