Be less aggressive about copying over canvas styles
When inlining canvases, we were calling getComputedStyle to get all the
styles from the canvas over to the image. That worked well if the screen
size is static, but on responsive setups that becomes a problem. Since
getComputedStyle contains width and height, the inlined image will be
statically sized after the canvas at the size it was rendered in the
cypress test, which might be a larger or smaller size than what it would
render as in other screen sizes.