Include externally referenced css content in asset package
In some cases, the externally downloaded css content has references to
local resources. If we want things to look right, we need to include
these in the assets package.

I made some small improvements in a few places while I was here. One of
them included generating a hash for the filename instead of uri
decoding. This made asset resolution on workers more stable.