Avoid using `crypto` module in browser
This commit addresses #31. That issue describes this error being thrown
when upgrading to Cypress v5:

```
crypto.createHash is not a function
```

While I haven't been able to reproduce this issue myself, it's a little
surprising that the crypto module is available in the browser that
Cypress runs. I'm switching to a browser-safe md5 implementation
instead, which I'm hoping will resolve the issue.