Fix md5 hash creation for style deduplication
In 9395de0, I replaced the usage of the built-in crypto module with a
browser-safe alternative. To test out the new module (crypto-js) I
relied on inlined canvases, which use md5 hashing to produce a file
name. That worked well, but I was also using md5 hashing for style
deduplication, and accidentally introduced a bug that caused style
elements to be processed multiple times.

Crypto-js docs are a little sparse on the return value of `md5()`.
https://github.com/brix/crypto-js

I assumed it was a string, partially convinced by the fact that inlined
canvas images ended up with a hash in their names. But the return value
is actually an object, you need to call `toString()` to get the string
version of the hash. For inlined canvases, this happened implicitly,
`/_inlined/${md5(svg)}.png`. For styles, we were using the object which
wasn't working (but didn't show up in our tests).

Following up on this commit, I'm going to revert both 45c58b3 and
386e0a5, because I believe hashing should be enough to avoid duplicate
styles.