Ignore assets without urls
I'm debugging a case where an asset is using an empty string as the src
url. The asset package fails somewhat silently with an error log and the
happoTeardown call times out.

We can ignore these assets as they won't work anyway. While I was at it,
I made the createAssetPackage function reject when an archiver error
happens. This will help users as the error will be propagated to the
Cypress process.