Fix making external asset urls absolute
When we discover an external CSS file, we download it and send it down
as `globalCSS`. An important step before we can inject into `globalCSS`
is that we "externalize" all urls found in the CSS file. E.g.

  url(/foo/bar.png);

needs to become

  url(https://base.url/foo/bar.png);

otherwise the asset won't be found on the happo workers.