Reset knownComponentVariants when report is created
This fixes a bug where if you use `cypress open` and run the same test
twice in a row you'll end up with component variants being incorrect:

default-2 instead of default
large-2 instead of large
etc

In 711eb197, I removed the reset of the knownComponentVariants variable
since it messed up reports when using the happo-cypress wrapper. That
commit introduced the issue fixed here. We have to be careful about
resetting and only do it in the non-wrapper case, otherwise we might end
up with invalid reports.