Work around performance issues (timeouts) with happoRegisterSnapshot
We are currently seeing the `happoRegisterSnapshot` call time out in
certain test suites. It seems related to the size of the payloads sent
over the Cypress websocket -- we've only seen this with pages containing
large canvases (that get inlined as base64 images).

I was led to this solution after finding this Cypress issue:
https://github.com/cypress-io/code-coverage/issues/76

In that thread, it's suggested that you bypass the built-in
serialization and deserialization in Cypress by using JSON.stringify
yourself.