Send inlined canvases separately from snapshot
This commit will hopefully resolve a timeout issue with
happoRegisterSnapshot (most likely) related to large snapshots with
inlined canvases. Instead of keeping the base64 url in the assets list,
we can eagerly send it in a new happoRegisterBase64Image task, store it
to disk and let the asset package simply pull in the image on disk. This
should help keep the main happoRegisterSnapshot payload smaller, which
I'm hoping will make it succeed more often. It also makes the in-memory
footprint smaller since the base64 images are discarded sooner.

For now I'm keeping a .happo-tmp folder around. I might make an effort
to automatically clean it up at some point in the future but I want to
make sure this fix is the right one first.