Add responsiveInlinedCanvases option
This new option will control the behavior of the inlined png placed
where canvases are found. The default is to copy over as much styling as
possible from the canvas element, including width and height attributes.
We've found cases where this is problematic however, and where it is
causing things to look off (since the canvas has a fixed width and
height).

By introducing a new `responsiveInlinedCanvases` option, we can allow
users to control the behavior better. There's a global way to configure
things in commands.js as well as a local option in happoScreenshot.

// global
import { configure } from 'happo-cypress';
configure({ responsiveInlinedCanvases: true });

// local
cy.get('selector').happoScreenshot({ responsiveInlinedCanvases: true });

The default is `false`, which is the old behavior.