Don't include data uris in asset packages
Yesterday I was debugging issues on our happo-snap workers. For some
reason, asset packages were containing entries where the file name was a
data uri. I tracked this down to the happo-cypress module. If an image
is using a data-uri it would be included in the list of urls sent to
createAssetsPackage. If the request made to download the image resulted
in a 200 response, that content would be included in the assets package.

I fixed this issue on our workers by inspecting the zip package and only
extracting ones that we know are okay.

By filtering out data uris before we send them to the happo task, we can
get rid of this problem from the source.