Disable Happo by default when running `cypress open`
A customer reached out and was confused about Happo running for each
time they ran tests locally. This is quite a gotcha, and can lead to
quota/cost issues down the line.

The right thing to do is to disable Happo by default and only run it
when explicitly asked for (via the wrapper). I thought about adding a
configuration option somewhere to enable the old behavior but decided an
env variable was enough, as this is more of a per-run setting.