Add support for dynamic targets
To make it easier to map the Happo browser targets to the viewports used
in the Cypress tests, we're making it possible to inject Happo targets
at runtime.

We're piggy-backing on the `targets` option here. If the target is a
string, use as before (limiting the screenshot to that target). If it is
an object, we merge it with the current .happo.js targets, then limit to
the new target.

Being able to dynamically add targets at runtime will make the Cypress
run better coupled to the Happo screenshots, and there's less risk for
them getting out of sync. Often times, dynamic targets will be
constructed automatically from Cypress runtime properties.