Only use dynamically added target when asked for
I found a bug with the previous commit that introduced the concept of a
dynamic target. Since we modify the happoConfig.targets property,
happoSnapshots taken after the first dynamic one will also include the
dynamically added target. Which is not what we want. To help resolve
this, I'm adding a filter on targets even when it's not set for the
`happoScreenshot`. If it's empty/undefined, we will only use the
non-dynamic targets.