Add example of a big canvas
We're seeing the happoRegisterBase64Image task time out for an
account. Their canvases are quite big, so I suspect there's a oom issue
of some sort here. Going to add a large canvas to our test suite to see
if it causes CI to fail the same way for us.