Chunk up saving the base64 image
The total size of the base64 package sent to the
happoRegisterBase64Image can be quite large, and we're seeing occasional
task timeouts (most likely because the task process runs out of memory
and exits).

To mitigate this, I'm sending the base64 image in chunks.  Each chunk is
appended to a file in the file system, and as a last step the base64
file is converted to a real file (using streams to prevent reading large
things into memory).