Resolve github actions environments automatically
A couple of recent customers have asked questions about how to set
things up in github actions. In this commit, I'm duplicating some of the
logic in the happo-ci-github-actions script for the happo.io module:
https://github.com/happo/happo.io/blob/073a884f0ac4edb7d6e641918b330742f5909544/bin/happo-ci-github-actions

We need to handle push and pull_request events. Compared to other CI
systems, we get fewer environment variables. Instead, we have to grab
variables from the associated event, linked to by the
`GITHUB_EVENT_PATH` env variable.