Fix resolving relative assets in non-root documents
Our previous use of doc.location.origin as the base url caused issues
when you were on a page with a path, e.g.

http://host.domain/foo/index.html

and had a relative path to a stylesheet, e.g.

themes/bar/styles.css

We ended up resolving the path to the css as
http://host.domain/themes/bar/styles.css, which is incorrect -- we need
the path involved in the base url as well.