Associate snap-requests with report sooner
When there is a nonce, meaning we're in parallel run mode, there is a
risk the last step of the happo-cypress wrapper is lost. Might be from a
failing test or some internal hickup. In these cases, it would make
sense to associate the snap-requests with the report sooner.

This can help fix issues for certain accounts seeing some reports with
missing snapshots. I've investigated and found that in some cases there
are orphaned snap-requests right around the time for when a report was
constructed (that had missing snapshots).