Find document from subject element
This commit fixes an issue where taking a screenshot of an iframe body
would result in no styles being present in the screenshot. This is
because we were assuming that all elements and styles were present in
the root cypress document. By grabbing the document relative to the
element we're working with, we can make screenshots of iframe content
look better.

Please note that this doesn't fix the issue where taking a screenshot of
some element that wraps an iframe results in the iframe being empty.
This commit will only make it so that you can dive into an iframe and
select elements (including <body>) inside.