Inject classes from <html> element before screenshot
I've seen classes on the `html` element affect styling a few times now,
and I decided to do something about this. By grabbing the html classes
at happoScreenshot time, and then injecting them at screenshot time, we
can get better style matching with the "real" website.

An example of a style I'm testing this on is a website with a
`fonts-loaded` class set. This class affects what fonts are shown in the
UI.