Remove retried snapshots before sending to happo.io
If a test is automatically retried, happo would duplicate snapshots that
were part of the retry and add a counter suffix to the variant name,
e.g. "foo-bar-1".

To fix this, I've made us listen to the after:spec event, then inspect
each test to look for failed attempts. If a failed attempt is found, we
delete all snapshots made during the duration of the failed attempt. I'm
making the assumption that timestamps will be in sync between the
browser process and the background node process. This could be
dangerous, but I couldn't find a better alternative to associate
snapshots with a test attempt.

Because of the added after:snapshot event handler, I'm changing the
initialization process:

happoTask.register(on)

This is a breaking change, so in order to help people upgrading without
updating config, I'm adding an assertion and an error message if we
detect that we haven't registered correctly.