Fix running with `cypress open`
When in cypress open mode (interactive), we need to set
experimentalInteractiveRunEvents=true in config to have before:spec and
after:spec events to fire. Once opting in, we were getting an error on
after:spec because in interactive mode, the after:spec isn't called with
any results. We can simply presence-check things to avoid the error.