Scroll home_page_spec.js up and down to begin with
This will ensure that any lazy-loaded images are loaded. In our case, we
have an image loaded via next/image. On initial render, this image is a
blank 1x1px gix image. Then when the image is in view, the real source
is loaded and injected. To get happo screenshots containing this image,
we need to make sure the lazy event happens. So we scroll down the page
quickly (and then back up).

Using a duration ensured that each section of the page came into view
(if only for a brief moment). First I tried using scrollTo without a
duration, and the lazy loading didn't happen then.