Fix removing duplicate screenshots for Cypress 13
In Cypress 13, each `attempt` in the spec results no longer have start
time and duration. Instead it simply has a `state` property. We rely on
time frame information for specs that fail, so that we can clean out
duplicate happo screenshots taken during the multiple attempts. To work
around this, I found that Cypress takes a screenshot image of the failed
state, so I'm using the timestamp of this image to construct the cleanup
timeframe. Probably brittle, but works for now.