Prevent allow-failures tests from failing CI jobs
We have some tests here that exercise the allow-failures flag which was
added here:

https://github.com/happo/happo-cypress/pull/36

This flag will allow Happo to finalize the report even if there are
Cypress failures, but will still exit non-zero if there are Cypress
failures.

This currently causes our CI jobs to fail, since we introduce
intentional Cypress failures. But this is actually the opposite of what
we want to test. Instead, we want the job to fail if the exit code is 0,
since we are expecting it to be not zero.