Execute applyConstructedStylesPatch before window loads
We are working to fix an issue where some constructed styles are not
being correctly sent to Happo for screenshots. We started this fix here:

https://github.com/happo/happo-e2e/pull/75

In Cypress, we need to use a slightly different approach to what we do
in that change--we need to pass it a reference to the window since
Cypress runs the document in an iframe. We added a property
takeDOMSnapshot that has the function we can call and pass in the window
reference here:

https://github.com/happo/happo-e2e/pull/76